Final Product - Fixed Responsiveness & Completed TODO List
The header is now sticky and will flow with the page.The site will
wrap as the window decreases making it easy for mobile users to enjoy.
This is applied to the top, mid-top, and mid-bottom sections. Max-width
was applied in order to keep the page centers for large screens (I have
an ultrawide).

Padding, margins, and some flex additons to make the site appear
cleaner. Stylized button cursors and made the header li an anchor.